refactor: enable script to be used to delete single accounts
